feat: Add API endpoint for voting on posts
This commit adds a new API endpoint in the `votes.py` file to handle voting on posts. The endpoint allows users to vote on a post by sending a POST request with the necessary parameters. If the vote is successful, a message is returned indicating the success. If the user has already voted on the post, a 400 Bad Request error is returned. If the user wants to remove their vote, they can send a POST request with the `status` parameter set to `False`. If the vote does not exist, a 404 Not Found error is returned.

This change enhances the functionality of the application by allowing users to interact with posts through voting, providing a more engaging user experience.